requirements.txt
ipykernel
pandas
datasets
seaborn
kagglehub
scikit-learn
matplotlib
torch
transformers
fastapi
uvicorn
streamlit
pinecone
sentence-transformers
openai
google-genai